requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mistralai>=1.2.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.27.0